+++ test/validation/sensitive/valid.ts
@@ -1,0 +1,5 @@
+import { sensitive, string } from '../../../src';
+
+export const Struct = sensitive(string());
+export const data = 'hello';
+export const output = 'hello';